+++ jme3-android/src/main/java/com/jme3/app/state/VideoRecorderAppState.java
@@ -358,7 +358,7 @@ public class VideoRecorderAppState extends AbstractAppState {
 
         @Override
         public float getTimePerFrame() {
-            return (float) (1.0f / this.framerate);
+            return 1.0f / this.framerate;
         }
 
         @Override
