+++ jme3-terrain/src/main/java/com/jme3/terrain/heightmap/HillHeightMap.java
@@ -135,7 +135,7 @@ public class HillHeightMap extends AbstractHeightMap {
         // transfer temporary buffer to final heightmap
         for (int i = 0; i < size; i++) {
             for (int j = 0; j < size; j++) {
-                setHeightAtPoint((float) tempBuffer[i][j], j, i);
+                setHeightAtPoint(tempBuffer[i][j], j, i);
             }
         }
