+++ jme3-vr/src/main/java/com/jme3/post/CartoonSSAO.java
@@ -140,7 +140,7 @@ public class CartoonSSAO extends Filter{
         frustumNearFar = new Vector2f();
 
         float farY = (vp.getCamera().getFrustumTop() / vp.getCamera().getFrustumNear()) * vp.getCamera().getFrustumFar();
-        float farX = farY * ((float) screenWidth / (float) screenHeight);
+        float farX = farY * (screenWidth / (float) screenHeight);
         frustumCorner = new Vector3f(farX, farY, vp.getCamera().getFrustumFar());
         frustumNearFar.x = vp.getCamera().getFrustumNear();
         frustumNearFar.y = vp.getCamera().getFrustumFar();
